requirements.txt
Flask
Werkzeug
opencv-python
numpy==1.24.4
Pillow
torch
torchvision
paddleocr
paddlepaddle
spacy
fuzzywuzzy
python-Levenshtein
scikit-image
fpdf
transformers
sentencepiece
sacremoses